+++ src/wasm-builder.h
@@ -239,11 +239,21 @@ public:
   }
   Break* makeBreak(Name name,
                    Expression* value = nullptr,
-                   Expression* condition = nullptr) {
+                   Expression* condition = nullptr,
+                   std::optional<Type> type = std::nullopt) {
     auto* ret = wasm.allocator.alloc<Break>();
     ret->name = name;
     ret->value = value;
     ret->condition = condition;
+    if (value && condition) {
+      // This is a br_if, for whom we must receive the type, as the type must
+      // match the block it targets (which we do not know here).
+      //
+      // Note that we could infer the type in some cases, like for MVP types,
+      // but we intentionally do not so as to catch bugs earlier.
+      assert(type);
+      ret->type = *type;
+    }
     ret->finalize();
     return ret;
   }